fix: separate and cleanup tests
